Update xah-fly-keys.el
buxfix: 


The idea is to make 【a】 call execute-extended-command or smex or helm (if they are installed). But this is something that is only supposed to happen when in command mode. And not in insert mode. Thus, I moved the function. Otherwise you can't type the letter 【a】in insert mode. In other words, this makes 'insert mode' implementation stop and become essentially useless. Critical -- this also breaks minibuffer entry of commands or input arguments to commands, including any file names, buffer names, Emacs command names, or Lisp expressions containing the letter 【a】.